% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ban_geocode_}
\alias{ban_geocode_}
\alias{ban_geocode}
\title{geocoding using the French national address database (BAN)
http://adresse.data.gouv.fr/}
\usage{
ban_geocode_(data, adresses, code_insee = NULL, code_postal = NULL,
  URL = "http://api-adresse.data.gouv.fr/")

ban_geocode(data, adresses, code_insee = NULL, code_postal = NULL,
  URL = "http://api-adresse.data.gouv.fr/")
}
\arguments{
\item{data}{a \code{dataframe} including at least one variable with addresses.}

\item{adresses}{name of the address column to be geocoded}

\item{code_insee}{name of the column including the French national code for a commune. If \code{NULL} (the default), this parameter is not used.}

\item{code_postal}{name of the column including  postal code. If \code{NULL} (the default), this parameter is not used. If \code{code_insee} and \code{code_postal} are provided, only \code{code_insee} is used.}

\item{URL}{base url of the API. This should only be modified if another endpoint than the one provided by Etalab (\url{http://api-adresse.data.gouv.fr/}) is used.}
}
\value{
an augmented dataframe with columns about geolocalization
}
\description{
geocoding using the French national address database (BAN)
http://adresse.data.gouv.fr/
}
\details{
the function \code{ban_geocode_} use a formula for \code{adresses}, when the function \code{ban_geocode} works with column name without quote in the `dplyr` style.
}
\examples{
library(dplyr)
data(paris2012)
test_df <- paris2012 \%>\%
             slice(1:10) \%>\%
             mutate(adresse = paste(numero, voie, nom),
                    code_insee = paste0("751", arrondissement))
ban_geocode(test_df, adresse, code_insee = "code_insee")

}
