% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply a function to columns of a data frame satisfying a condition}
\usage{
df_apply(.data, .f, .condition = is.numeric, .else = identity, ...)
}
\arguments{
\item{.data}{a data frame.}

\item{.f}{a function to apply to columns of the data frame.}

\item{.condition}{a function returning boolean values.}

\item{.else}{a function to apply when .condition evaluates to false.}

\item{...}{additional arguments to be passed to .f.}
}
\value{
A tibble with updated columns.
}
\description{
Apply a function to columns of a data frame satisfying a condition
}
\examples{

}
