% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{F1racing}
\alias{F1racing}
\title{Statuses of Formula 1 races}
\format{
A data frame with 139 rows and 3 variables:
\describe{
\item{status}{finished or accident/mechanical failure}
\item{status_sum}{count of the statuses for a given year}
...
}
}
\source{
\url{http://ergast.com/mrd/db/#csv/}
}
\usage{
F1racing
}
\description{
A dataset containing attributes of Formula 1 races over the years.
}
\keyword{datasets}
