% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_consumption_grouped}
\alias{food_consumption_grouped}
\title{Food Data of Animal and non-Animal Products}
\format{
A data frame with 2 rows and 6 variables:
\describe{
\item{food_type}{Animal Product or non-Animal Product}
\item{total_consumption}{global consumption, in kg/person/year}
\item{total_emission}{global CO2 emission, in kg/person/year}
\item{perc_consumption}{percentage of total global food consumption}
\item{perc_emission}{percentage of total global food CO2 emission}
\item{\% of Total}{placeholder for graph label}
...
}
}
\source{
\url{https://www.nu3.de/blogs/nutrition/food-carbon-footprint-index-2018}
}
\usage{
food_consumption_grouped
}
\description{
A dataset containing the global consumption and carbon emissions of Animal
and non-Animal products.
}
\keyword{datasets}
