% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_consumption_total}
\alias{food_consumption_total}
\title{Food Data of All Food Groups}
\format{
A data frame with 11 rows and 4 variables:
\describe{
\item{food_category}{food group}
\item{food_type}{Animal Product or non-Animal Product}
\item{total_consumption}{global consumption, in kg/person/year}
\item{total_emission}{global CO2 emission, in kg/person/year}
...
}
}
\source{
\url{https://www.nu3.de/blogs/nutrition/food-carbon-footprint-index-2018}
}
\usage{
food_consumption_total
}
\description{
A dataset containing the global consumption and carbon emissions of
different Food Groups
}
\keyword{datasets}
