% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_object_size}
\alias{format_object_size}
\title{Format object size}
\usage{
format_object_size(x, units = "b", standard = "auto", digits = 1L, ...)
}
\arguments{
\item{x}{a number}

\item{units}{a unit}

\item{standard}{a standard}

\item{digits}{number of digits}

\item{...}{anything else}
}
\description{
This function is modified copy of the utils:::format.object_size 
function which is not exported. 
The main difference is that it returns values.
}
