% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode}
\usage{
geocode(
  query,
  limit = NULL,
  autocomplete = NULL,
  type = c("all", "housenumber", "street", "locality", "municipality")
)
}
\arguments{
\item{query}{a string of the address you want to geocode}

\item{limit}{an integer indicating the maximum number of results wanted}

\item{autocomplete}{a boolean}

\item{type}{a string with the type of result wanted. 
Default returns all results. 
Can have one of the following values: "all", "housenumber", "street", "locality" or "municipality"}
}
\value{
a tibble
}
\description{
Geocode
}
\examples{
geocode(query = "39 quai André Citroën, Paris")

}
