% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_csv.R
\name{geocode_csv}
\alias{geocode_csv}
\title{Geocode French addresses using a CSV file}
\usage{
geocode_csv(file, ...)
}
\arguments{
\item{file}{the path of the csv file containing the addresses.}

\item{...}{additional arguments provided by the API.}
}
\value{
a tibble with geocoded addresses among other related information
}
\description{
Geocode French addresses using a CSV file
}
\examples{

# Downloading the example csv file provided in the API website

download.file("https://adresse.data.gouv.fr/exemples/search.csv", "search.csv")

# using this csv file

geocode_csv("search.csv")



}
