% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_csv_reverse.R
\name{geocode_csv_reverse}
\alias{geocode_csv_reverse}
\title{Inverse geocoding geographical coordinates using a csv file}
\usage{
geocode_csv_reverse(file, ...)
}
\arguments{
\item{file}{the path of the csv file containing the coordinates}

\item{...}{additional arguments provided by the API.}
}
\value{
a tibble with reverse geocoded coordinates among other related information
}
\description{
Inverse geocoding geographical coordinates using a csv file
}
\examples{

# Downloading the example csv file provided in the API website

download.file("https://adresse.data.gouv.fr/exemples/reverse.csv", "reverse.csv")

# using this csv file

geocode_csv_reverse("reverse.csv")

}
