% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_df.R
\name{geocode_df}
\alias{geocode_df}
\title{Geocode df}
\usage{
geocode_df(data, adresses, code_insee = NULL, code_postal = NULL)
}
\arguments{
\item{data}{}

\item{adresses}{}

\item{code_insee}{}

\item{code_postal}{}
}
\value{
a tibble
}
\description{
Geocode df
}
\examples{

\dontrun{
geocode_df(data = table_test, adresses = adresse)
}

}
