% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_df.R
\name{geocode_df_}
\alias{geocode_df_}
\title{Geocode df}
\usage{
geocode_df_(data, adresses, code_insee)
}
\arguments{
\item{data}{}

\item{adresses}{}

\item{code_insee}{}
}
\value{
a tibble
}
\description{
Geocode df
}
\examples{

\dontrun{
geocode_df_(data = table_test, adresses = "adresse")
}

}
