% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_tbl.R
\name{geocode_tbl}
\alias{geocode_tbl}
\title{Geocode tbl}
\usage{
geocode_tbl(tbl, adresse, code_insee = NULL, code_postal = NULL)
}
\arguments{
\item{tbl}{a data frame or tibble}

\item{adresse}{adress column}

\item{code_insee}{official citycode column}

\item{code_postal}{official postcode column}
}
\value{
an augmented data frame of classe tbl with latitude, longitude, etc
}
\description{
Geocode tbl geocodes a whole data frame
}
\examples{

table_test <- tibble::tibble(
x = c("39 quai André Citroën", "64 Allée de Sully", "20 avenue de Ségur"), 
y = c("75015", "75012", "75007"), 
z = rnorm(3)
)

geocode_tbl(tbl = table_test, adresse = x)
geocode_tbl(tbl = table_test, adresse = x, code_postal = y)

}
