% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{get_features}
\alias{get_features}
\title{Get features}
\usage{
get_features(x)
}
\arguments{
\item{x}{the content of a request}
}
\value{
a tibble
}
\description{
Get features
}
\examples{
library("magrittr")
library("httr")
library("tibble")
library("purrr")
test <- "http://api-adresse.data.gouv.fr/search/?q=8 bd du port" \%>\% 
URLencode() \%>\% 
GET() \%>\% 
content()
get_features(test)

}
