% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{get_geometry}
\alias{get_geometry}
\title{Get geometry}
\usage{
get_geometry(x)
}
\arguments{
\item{x}{a feature}
}
\value{
a tibble
}
\description{
Get geometry
}
\examples{
library("magrittr")
library("httr")
library("tibble")
library("purrr")
test <- "http://api-adresse.data.gouv.fr/search/?q=8 bd du port" \%>\% 
URLencode() \%>\% 
GET() \%>\% 
content()
feature <- extract2(test, "features")[[1]]
get_geometry(feature)
map_df(.x = test$features, .f = get_geometry)

}
