% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_df.R
\name{post_request}
\alias{post_request}
\title{Post Request}
\usage{
post_request(tempfile, adresse, code_insee = NULL, code_postal = NULL)
}
\arguments{
\item{code_insee}{}

\item{code_postal}{}

\item{data}{}

\item{adresses}{}
}
\value{
results of a request
}
\description{
Post Request
}
\examples{

table_test <- tibble::tibble(
x = c("39 quai André Citroën", "64 Allée de Sully", "20 avenue de Ségur"), 
y = c("75015", "75012", "75007"), 
z = rnorm(3)
)

tmp <- write_tempfile(data = table_test, adresse = x)
post_request(
tempfile = tmp, 
adresse = x
)
 
}
