% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Reverse geocode}
\usage{
reverse_geocode(long, lat)
}
\arguments{
\item{long}{longitude}

\item{lat}{latitude}
}
\value{
a tibble
}
\description{
Reverse geocode
}
\examples{
\dontrun{
reverse_geocode(long = 2.37, lat = 48.357)
}

}
