% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_tbl.R
\name{reverse_geocode_tbl}
\alias{reverse_geocode_tbl}
\title{Reverse geocode tbl}
\usage{
reverse_geocode_tbl(tbl, longitude, latitude)
}
\arguments{
\item{tbl}{name of the tibble}

\item{longitude}{name of the longitude column}

\item{latitude}{name of the latitude column}
}
\value{
an augmented tibble with addresses
}
\description{
reverse geocode a data frame
}
\examples{
\dontrun{
table_reverse <- tibble::tibble(
x = c(2.279092, 2.375933,2.308332),
y = c(48.84683, 48.84255, 48.85032),
z = rnorm(3)
)

reverse_geocode_tbl(tbl = table_reverse, longitude = x, latitude = y)
}

}
