% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USA_Annual_GDP}
\alias{USA_Annual_GDP}
\title{Annual GDP data for the USA.}
\format{
A data frame with 92 rows and 31 variables:
\describe{
\item{Year}{year}
\item{GDP}{total nominal GDP, in millions of dollars}
\item{Consumption}{total spent on goods/services, in millions of dollars}
\item{Investment}{total spent on investment, in millions of dollars}
\item{Government.Spending}{total spent by government, in millions of dollars}
\item{Net.Exports}{total exports minus total imports, in millions of dollars}
\item{Consumption.prop}{Consumption / GDP}
\item{Investment.prop}{Investment / GDP}
\item{Government.Spending.prop}{Government.Spending / GDP}
\item{Net.Exports.prop}{Net.Exports / GDP}
...
}
}
\source{
\url{https://apps.bea.gov/iTable/iTable.cfm?reqid=19&step=2}
}
\usage{
USA_Annual_GDP
}
\description{
A dataset containing the annual nominal GDP for the USA, divided up into
many categories (1929-2020).
}
\keyword{datasets}
