% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_df.R
\name{write_tempfile}
\alias{write_tempfile}
\title{Write tempfile}
\usage{
write_tempfile(data, adresse, code_insee = NULL, code_postal = NULL)
}
\arguments{
\item{data}{a data frame}

\item{code_insee}{name of insee code column}

\item{code_postal}{name of the postal code column}

\item{adresses}{name of the address column}
}
\value{
path to the temp file
}
\description{
Write tempfile
}
\examples{
table_test <- tibble::tibble(
x = c("39 quai André Citroën", "64 Allée de Sully", "20 avenue de Ségur"), 
y = c("75015", "75012", "75007"), 
z = rnorm(3)
)
write_tempfile(data = table_test, adresse = x)

}
