% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ban_geocode_}
\alias{ban_geocode}
\alias{ban_geocode_}
\title{géocode des adresses grâce à l'API BANO}
\usage{
ban_geocode_(data, adresses, code_insee = NULL, code_postal = NULL,
  URL = "http://api-adresse.data.gouv.fr/")

ban_geocode(data, adresses, code_insee = NULL, code_postal = NULL,
  URL = "http://api-adresse.data.gouv.fr/")
}
\arguments{
\item{data}{un \code{dataframe} contenant, a minima, les adresses à géocoder.}

\item{adresses}{nom de la colonne contenant les adresses à géocoder.}

\item{code_insee}{nom d'une colonne contenant le code INSEE de la commune. Si \code{NULL} (par défaut), ce paramètre n'est pas utilisé.}

\item{code_postal}{nom d'une colonne contenant le code postal de la commune. Si \code{NULL} (par défaut), ce paramètre n'est pas utilisé. Si \code{code_insee} et \code{code_postal} sont tous les deux fournis, seul \code{code_insee} est fourni.}

\item{URL}{base de l'URL de l'API à utiliser. Ne modifier qu'en cas de déploiement local ou sur un serveur distant de l'API BAN, distinct du point d'entrée offert par Etalab (\url{http://api-adresse.data.gouv.fr/}).}
}
\value{
un dataframe avec des colonnes supplémentaires avec la géolocalisation
}
\description{
géocode des adresses grâce à l'API BANO
}
\details{
la fonction \code{ban_geocode_} utilise une formule pour l'argument \code{adresses}, tandis que la fonction \code{ban_geocode} utilise un nom de colonne sans guillemets, à la dplyr.
}
\examples{
library(dplyr)
data(paris2012)
test_df <- paris2012 \%>\%
             slice(1:10) \%>\%
             mutate(adresse = paste(numero, voie, nom),
                    code_insee = paste0("751", arrondissement))
ban_geocode(test_df, adresse, code_insee = "code_insee")

}

