% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_df.R
\name{geocode_df}
\alias{geocode_df}
\title{Geocode df
 
Function to geocode a data frame}
\usage{
geocode_df(data, adresse, code_insee = NULL, code_postal = NULL)
}
\arguments{
\item{data}{a dataframe}

\item{adresse}{column including adresses}

\item{code_insee}{column including code insee}

\item{code_postal}{column including post code}
}
\value{
a tibble
}
\description{
Geocode df
 
Function to geocode a data frame
}
\examples{

table_test <- tibble::tibble(
x = c("39 quai André Citroën", "64 Allée de Sully", "20 avenue de Ségur"), 
y = c("75015", "75012", "75007"), 
z = rnorm(3)
)

geocode_df(data = table_test, adresse = x)
geocode_df(data = table_test, adresse = x, code_postal = y)

}
