% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paris2012}
\alias{paris2012}
\title{Addresses in the electoral register of Paris, 2012.}
\format{
a \code{tbl_df} with 72107 lines and 7 variables
\describe{
 \item{arrondissement}{code of the arrondissement (district)}
 \item{bureau}{code of the polling station, in the arrondissement}
 \item{numero}{street number}
 \item{voie}{type of street}
 \item{nom}{name of the street}
 \item{nb}{number of voters registered at this adress}
 \item{ID}{polling station ID}
}
}
\source{
data have been collected by Baptiste Coulmont
}
\usage{
paris2012
}
\description{
This dataset includes addresses found in the Parisian electoral 
register in 2012.
}
\keyword{datasets}
