% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paris2012}
\alias{paris2012}
\title{Addresses in the electoral register of Paris, 2012.}
\format{un \code{tbl_df} avec 72107 lignes et 7 variables
\describe{
 \item{arrondissement}{code de l'arrondissement}
 \item{bureau}{code du bureau de vote, dans l'arrondissement}
 \item{numero}{numéro dans la voie}
 \item{voie}{type de voie}
 \item{nom}{nom de la voie}
 \item{nb}{nombre d'électeurs enregistrés à cette adresse}
 \item{ID}{identifiant du bureau de vote}
}}
\source{
data have been collected by Baptiste Coulmont
}
\usage{
paris2012
}
\description{
This dataset includes addresses found in the Parisian electoral register in 2012.
}
\keyword{datasets}

