% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_df.R
\name{post_request}
\alias{post_request}
\title{Post Request}
\usage{
post_request(data, adresses, code_insee = NULL, code_postal = NULL)
}
\arguments{
\item{data}{}

\item{adresses}{}

\item{code_insee}{}

\item{code_postal}{}
}
\value{
results of a request
}
\description{
Post Request
}
\examples{
\dontrun{
post_request(
data = "/tmp/RtmpQaOPOu/file2d0b69b2dc4a.csv", 
adresses = "adresse", 
code_insee = "code_insee")
results <- post_request(
data = "/tmp/RtmpQaOPOu/file2d0b69b2dc4a.csv", 
adresses = "adresse", 
code_insee = "code_insee")
}

}
