% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USA_Annual_GDP}
\alias{USA_Annual_GDP}
\title{Annual GDP data for the USA.}
\format{
A data frame with 92 rows and 31 variables:
\describe{
\item{Year}{year}
\item{GDP}{total nominal GDP, in millions of dollars}
\item{Consumption}{total spent on goods/services, in millions of dollars}
\item{Investment}{total spent on investment, in millions of dollars}
\item{Government.Spending}{total spent by government, in millions of dollars}
\item{Net.Exports}{total exports minus total imports, in millions of dollars}
\item{Consumption.prop}{Consumption / GDP}
\item{Investment.prop}{Investment / GDP}
\item{Government.Spending.prop}{Government.Spending / GDP}
\item{Net.Exports.prop}{Net.Exports / GDP}
...
}
}
\source{
\url{https://apps.bea.gov/iTable/iTable.cfm?reqid=19&step=2}
}
\usage{
USA_Annual_GDP
}
\description{
A dataset containing the annual nominal GDP for the USA, divided up into
many categories (1929-2020).
}
\examples{

library(plotly)
my_viz1 <- USA_Annual_GDP \%>\%
plot_ly() \%>\%
  add_trace(x = ~Year, y = ~GDP, type = "scatter", mode = "lines", name = "Nominal GDP\nBy Year",
            fill = "tozeroy", fillcolor = "rgba(168, 216, 234, 0.5)",
            hovertemplate = '<b>Year</b>: \%{x}<br><b>NGDP</b>: $\%{y}<br><i>(value in millions of USD)</i>') \%>\%
  layout(title = list(text = "Nominal GDP by Year", font = list(size = 22), y = 0.99),
         yaxis = list(title = "Nominal GDP (millions of USD)"),
         xaxis = list(rangeslider = list(type = "date")))

my_viz1
}
\keyword{datasets}
