% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_df.R
\name{write_tempfile}
\alias{write_tempfile}
\title{Write_temple}
\usage{
write_tempfile(data, adresses, code_insee)
}
\arguments{
\item{data}{a data frame}

\item{adresses}{name of the address column}

\item{code_insee}{name of code insee column}
}
\value{
path to the temp file
}
\description{
Write_temple
}
\examples{
\dontrun{
data(paris2012)
table_test <- paris2012 \%>\%
slice(1:10) \%>\%
mutate(
adresse = paste(numero, voie, nom), 
code_insee = paste0("751", arrondissement)
)
write_tempfile(data = table_test, adresses = "adresse", 
code_insee = "code_insee")
write_tempfile(data = table_test, adresses = "adresse")

temp_file <- write_tempfile(data = table_test, adresses = "adresse")
file.size(temp_file)
utils:::format.object_size(file.size(temp_file), "MB")
}

}
